+++ core/src/main/java/io/arcblock/forge/TransactionFactory.kt
@@ -6,7 +6,9 @@ import forge_abi.*
 import io.arcblock.forge.did.DIDGenerator
 import io.arcblock.forge.did.HashType
 import io.arcblock.forge.did.WalletInfo
-import io.arcblock.forge.extension.*
+import io.arcblock.forge.extension.delegatee
+import io.arcblock.forge.extension.multiSig
+import io.arcblock.forge.extension.toByteString
 import java.math.BigInteger
 import java.time.LocalDate
 import java.util.*
@@ -17,33 +19,34 @@ object TransactionFactory {
     return System.currentTimeMillis()
   }
 
+  const val POKE_ADDRESS = "zzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzz"
 
   /**
    * create a declare transaction without signature
    */
   @JvmOverloads
-  fun declare(chainID: String, wallet: WalletInfo, moniker: String? = null, issuer: String? = null): Type.Transaction {
+  fun declare(chainID: String, address: String, pk: ByteArray, moniker: String? = null, issuer: String? = null): Type.Transaction {
     val itx = Declare.DeclareTx.newBuilder()
       .setMoniker(moniker
         ?: UUID.randomUUID().toString().replace("-", ""))
-      .setIssuer(issuer ?: wallet.address.address())
+      .setIssuer(issuer ?: address)
       .build()
-    return createTransaction(chainID, wallet.address, wallet.pk, itx.toByteString(), TypeUrls.DECLARE)
+    return createTransaction(chainID, address, pk, itx.toByteString(), TypeUrls.DECLARE)
   }
 
 
   /**
    * create a poke transaction without signature
    */
-  fun unsignPoke(chainID: String, wallet: WalletInfo): Type.Transaction {
+  fun unsignPoke(chainID: String, address: String, pk: ByteArray): Type.Transaction {
     val itx = Poke.PokeTx.newBuilder()
-      .setAddress("zzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzz")
+      .setAddress(POKE_ADDRESS)
       .setDate(LocalDate.now().toString())
       .build()
     return Type.Transaction.newBuilder()
       .setChainId(chainID)
-      .setFrom(wallet.address)
-      .setPk(ByteString.copyFrom(wallet.pk))
+      .setFrom(address)
+      .setPk(ByteString.copyFrom(pk))
       .setNonce(0L)
       .setItx(Any.getDefaultInstance().toBuilder().setTypeUrl(TypeUrls.POKE)
         .setValue(itx.toByteString())
@@ -63,8 +66,10 @@ object TransactionFactory {
    *
    * @return Transaction
    */
-  fun unsignTransfer(chainID: String, from: String, fromPk: ByteArray,to: String, token: BigInteger? = null ,assets: List<String>? = null) :Type.Transaction {
-    val builder = Transfer.TransferTx.newBuilder().setTo(to)
+  @JvmOverloads
+  fun unsignTransfer(chainID: String, from: String, fromPk: ByteArray, to: String, token: BigInteger? = null, assets: List<String>? = null): Type.Transaction {
+    val builder = Transfer.TransferTx.newBuilder()
+      .setTo(to)
     token?.apply { builder.setValue(Type.BigUint.newBuilder().setValue(this.toByteArray().toByteString())) }
     assets?.apply { builder.addAllAssets(this) }
     return createTransaction(chainID, from, fromPk, builder.build().toByteString(), TypeUrls.TRANSFER)
@@ -82,24 +87,22 @@ object TransactionFactory {
    * @return Transaction
    */
   @JvmOverloads
-  fun preExchange(chainId: String, from: WalletInfo, to: String, fromToken: BigInteger, assetAddress: String, delegateeFrom: String? = null,
-                 delegateeTo: String? = null): Type.Transaction{
-      val exchange = Exchange.ExchangeTx.newBuilder()
-        .setSender(Exchange.ExchangeInfo.newBuilder()
-          .setValue(Type.BigUint.newBuilder()
-            .setValue(fromToken.toByteArray().toByteString())
-            .build())
+  fun preExchange(chainId: String, fromAddress: String, fromPk: ByteArray , to: String, fromToken: BigInteger, assetAddress: String, delegateeFrom: String? =
+    null): Type.Transaction {
+    val exchange = Exchange.ExchangeTx.newBuilder()
+      .setSender(Exchange.ExchangeInfo.newBuilder()
+        .setValue(Type.BigUint.newBuilder()
+          .setValue(fromToken.toByteArray().toByteString())
           .build())
-        .setReceiver(Exchange.ExchangeInfo.newBuilder()
-          .addAssets(assetAddress)
-          .build())
-        .setTo(delegateeTo ?: to)
-        .build()
-      return createTransaction(chainId, from.address, from.pk, exchange.toByteString(), TypeUrls.EXCHANGE)
-        .delegatee(delegateeFrom)
-        .signTx(from.sk)
-    }
-
+        .build())
+      .setReceiver(Exchange.ExchangeInfo.newBuilder()
+        .addAssets(assetAddress)
+        .build())
+      .setTo(to)
+      .build()
+    return createTransaction(chainId, fromAddress, fromPk, exchange.toByteString(), TypeUrls.EXCHANGE)
+      .delegatee(delegateeFrom)
+  }
 
 
   /**
@@ -113,30 +116,30 @@ object TransactionFactory {
    * @return Transaction
    */
   @JvmOverloads
-  fun preConusmeAsset(chainId: String, checker: WalletInfo, issuer: String? = null, delegatee: String? = null): Type.Transaction{
+  fun preUnsignConusmeAsset(chainId: String, checkerAddress: String, checkerPk: ByteArray, issuer: String? = null, delegatee: String? = null): Type
+  .Transaction {
     val consumeAsset = ConsumeAsset.ConsumeAssetTx.newBuilder()
-      .setIssuer(issuer ?: checker.address)
+      .setIssuer(issuer ?: checkerAddress)
       .build()
-    return createTransaction(chainId , checker.address, checker.pk, consumeAsset.toByteString(), TypeUrls.CONSUME_ASSET)
+    return createTransaction(chainId, checkerAddress, checkerPk, consumeAsset.toByteString(), TypeUrls.CONSUME_ASSET)
       .delegatee(delegatee)
-      .signTx(checker.sk)
   }
-
 
 
   /**
-   * finaize
+   * finalize  MultiSig
    */
-  fun finalizeMultiSig(tx: Type.Transaction, to: WalletInfo, delegateeTo: String? = null, data: Any? = null ): Type.Transaction{
+  @JvmOverloads
+  fun finalizeMultiSig(tx: Type.Transaction, to: WalletInfo, delegateeTo: String? = null, data: Any? = null): Type.Transaction {
     return tx.multiSig(to, delegator = delegateeTo, data = data)
   }
-
 
 
   /**
    * create a delegate transaction without signature
    */
-  fun unsignDelegate(from: String, to: String, chainID: String, wallet: WalletInfo, rules: List<String>, typeUrl: String? = null): Type.Transaction {
+  @JvmOverloads
+  fun unsignDelegate(from: String, to: String, chainID: String, fromPk: ByteArray, rules: List<String>, typeUrl: String? = null): Type.Transaction {
     val itx = Delegate.DelegateTx.newBuilder()
       .setAddress(DIDGenerator.genDelegateAddress(from, to))
       .addOps(Delegate.DelegateOp.newBuilder()
@@ -145,19 +148,19 @@ object TransactionFactory {
         .build())
       .setTo(to)
       .build()
-    return createTransaction(chainID, wallet.address, wallet.pk, itx.toByteString(), TypeUrls.DELEGATE)
+    return createTransaction(chainID, from, fromPk, itx.toByteString(), TypeUrls.DELEGATE)
   }
 
   /**
    * create a revoke delegate transaction without signature
    */
-  fun unsignRevokeDelegate(from: String, to: String, chainID: String, wallet: WalletInfo, typeUrls: List<String>): Type.Transaction {
+  fun unsignRevokeDelegate(from: String, to: String, chainID: String, fromPk: ByteArray, typeUrls: List<String>): Type.Transaction {
     val itx = RevokeDelegate.RevokeDelegateTx.newBuilder()
       .setAddress(DIDGenerator.genDelegateAddress(from, to))
       .addAllTypeUrls(typeUrls)
       .setTo(to)
       .build()
-    return createTransaction(chainID, wallet.address, wallet.pk, itx.toByteString(), TypeUrls.REVOKE_DELEGATE)
+    return createTransaction(chainID, from, fromPk, itx.toByteString(), TypeUrls.REVOKE_DELEGATE)
   }
 
   /**
@@ -175,13 +178,14 @@ object TransactionFactory {
       .build()
   }
 
-  fun setup_swap(chainId: String, from: WalletInfo, receiver: String, blockHeight: Int, hashKey: ByteArray, demandToken: BigInteger) =
-    setup_swap(chainId,from,receiver,blockHeight,hashKey,demandToken = demandToken, assets = listOf())
+  fun setupSwap(chainId: String, fromAddress:String, fromPk: ByteArray, receiver: String, blockHeight: Int, hashKey: ByteArray, demandToken: BigInteger) =
+    setupSwap(chainId, fromAddress, fromPk, receiver, blockHeight, hashKey, demandToken = demandToken, assets = listOf())
 
-  fun setup_swap(chainId: String, from: WalletInfo, receiver: String, blockHeight: Int, hashKey: ByteArray, assets: List<String>) =
-    setup_swap(chainId,from,receiver,blockHeight,hashKey,demandToken = null, assets = assets)
+  fun setupSwap(chainId: String, fromAddress:String, fromPk: ByteArray, receiver: String, blockHeight: Int, hashKey: ByteArray, assets: List<String>) =
+    setupSwap(chainId, fromAddress, fromPk, receiver, blockHeight, hashKey, demandToken = null, assets = assets)
 
-  fun setup_swap(chainId: String, from: WalletInfo, receiver: String, blockHeight: Int, hashKey: ByteArray, demandToken: BigInteger? = BigInteger.ZERO, assets:
+  fun setupSwap(chainId: String, fromAddress:String, fromPk: ByteArray, receiver: String, blockHeight: Int, hashKey: ByteArray, demandToken: BigInteger? = BigInteger.ZERO,
+                assets:
   List<String>? = listOf())
     : Type.Transaction {
     val itx = SetupSwap.SetupSwapTx.newBuilder()
@@ -193,22 +197,22 @@ object TransactionFactory {
       .setHashlock(Hasher.hash(HashType.SHA3, hashKey).toByteString())
       .setReceiver(receiver)
       .build()
-    return createTransaction(chainId, from.address, from.pk, itx.toByteString(), TypeUrls.SETUP_SWAP)
+    return createTransaction(chainId, fromAddress, fromPk, itx.toByteString(), TypeUrls.SETUP_SWAP)
   }
 
-  fun revoke_swap(chainId: String, from: WalletInfo, swapAddr: String): Type.Transaction {
+  fun revokeSwap(chainId: String, fromAddress:String, fromPk: ByteArray, swapAddr: String): Type.Transaction {
     val itx = RevokeSwap.RevokeSwapTx.newBuilder()
       .setAddress(swapAddr)
       .build()
-    return createTransaction(chainId, from.address, from.pk, itx.toByteString(), TypeUrls.REVOKE_SWAP)
+    return createTransaction(chainId, fromAddress, fromPk, itx.toByteString(), TypeUrls.REVOKE_SWAP)
   }
 
-  fun retrieve_swap(chainId: String, from: WalletInfo, swapAddr: String, hashKey: ByteArray): Type.Transaction {
+  fun retrieve_swap(chainId: String, fromAddress:String, fromPk: ByteArray, swapAddr: String, hashKey: ByteArray): Type.Transaction {
     val itx = RetrieveSwap.RetrieveSwapTx.newBuilder()
       .setAddress(swapAddr)
       .setHashkey(hashKey.toByteString())
       .build()
-    return createTransaction(chainId, from.address, from.pk, itx.toByteString(), TypeUrls.RETRIEVE_SWAP)
+    return createTransaction(chainId, fromAddress, fromPk, itx.toByteString(), TypeUrls.RETRIEVE_SWAP)
   }
 
 }